Fixed MeepMeep Library (#11)
